requirements.txt
#Python Version : 3.11.11
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-Limiter==3.5.0
pydantic==2.4.2
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7